flask_app/requirements.txt
Flask==3.0.3
matplotlib==3.8.2
numpy==1.24.3
Pillow==9.3.0
torch==2.1.2
transformers==4.36.2
